INS-1644 split to subfiles
