+++ ledger/storage/drop/db.go
@@ -18,66 +18,11 @@ package drop
 
 import (
 	"context"
-	"sync"
 
 	"github.com/insolar/insolar/core"
 	"github.com/insolar/insolar/ledger/storage"
 	"github.com/insolar/insolar/ledger/storage/jet"
 )
-
-// Modifier provides interface for modifying jetdrops
-//go:generate minimock -i github.com/insolar/insolar/ledger/storage/drop.Modifier -o ./ -s _mock.go
-type Modifier interface {
-	Set(ctx context.Context, jetID core.JetID, drop jet.Drop) error
-}
-
-// Accessor provides interface for accessing jetdrops
-//go:generate minimock -i github.com/insolar/insolar/ledger/storage/drop.Accessor -o ./ -s _mock.go
-type Accessor interface {
-	ForPulse(ctx context.Context, jetID core.JetID, pulse core.PulseNumber) (jet.Drop, error)
-}
-
-type dropKey struct {
-	pulse core.PulseNumber
-	jetID core.JetID
-}
-
-type dropStorageMemory struct {
-	lock sync.RWMutex
-	jets map[dropKey]jet.Drop
-}
-
-// NewStorageMemory creates new storage, that holds data in-memory
-func NewStorageMemory() *dropStorageMemory { // nolint: golint
-	return &dropStorageMemory{
-		jets: map[dropKey]jet.Drop{},
-	}
-}
-
-// ForPulse returns a jet.Drop for a provided pulse, that is stored in memory
-func (m *dropStorageMemory) ForPulse(ctx context.Context, jetID core.JetID, pulse core.PulseNumber) (jet.Drop, error) {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	key := dropKey{jetID: jetID, pulse: pulse}
-	d, ok := m.jets[key]
-	if !ok {
-		return jet.Drop{}, core.ErrNotFound
-	}
-
-	return d, nil
-}
-
-// Set saves a provided jet.Drop to memory
-func (m *dropStorageMemory) Set(ctx context.Context, jetID core.JetID, drop jet.Drop) error {
-	m.lock.RLock()
-	defer m.lock.RUnlock()
-
-	key := dropKey{jetID: jetID, pulse: drop.Pulse}
-	m.jets[key] = drop
-
-	return nil
-}
 
 type dropStorageDB struct {
 	DB storage.DBContext `inject:""`
